Updated ScoreKeeper and PlayerManager.players
I changed PlayerManager.players to be an array of Player Components rather than PlayerInput Components. This was for future proofing and it made more sense but this changed almost every script a little bit.

Also, ScoreKeeper's methods now take a Player as an argument rather than index, so the scoreKeeper manages all indexing rather than other scripts doing so.